requirements.txt
asgiref==3.5.0
attrs==21.4.0
autopep8==1.6.0
dj-database-url==0.5.0
Django==4.0.3
django-crispy-forms==1.14.0
django-extensions==3.1.5
django-filter==21.1
django-on-heroku==1.1.2
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
drf-spectacular==0.22.0
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.4.0
Pillow==9.1.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyrsistent==0.18.1
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
uritemplate==4.1.1
whitenoise==6.0.0
